Refactor state to a separate module
