+++ src/rust/ide/parser/src/lib.rs
@@ -73,10 +73,13 @@ impl Parser {
         Parser::new().unwrap_or_else(|e| panic!("Failed to create a parser: {:?}", e))
     }
 
+    /// Parse program.
     pub fn parse(&self, program:String, ids:IdMap) -> api::Result<Ast> {
         self.borrow_mut().parse(program,ids)
     }
 
+    /// Parse contents of the program source file, where program code may be followed by idmap and
+    /// metadata.
     pub fn parse_with_metadata<M:api::Metadata>
     (&self, program:String) -> api::Result<api::SourceFile<M>> {
         self.borrow_mut().parse_with_metadata(program)
