+++ src/rust/ide/src/controller/module.rs
@@ -5,177 +5,21 @@
 //! for registering text and graph changes. If for example text represntation will be changed, there
 //! will be notifications for both text change and graph change.
 
+pub mod state;
+
 use crate::prelude::*;
 
-use crate::controller::notification;
 use crate::double_representation::text::apply_code_change_to_id_map;
-use crate::double_representation::definition::DefinitionInfo;
-use crate::executor::global::spawn;
 
-use parser::api::SourceFile;
 use ast;
-use ast::Ast;
 use ast::HasRepr;
 use ast::HasIdMap;
 use data::text::*;
 use double_representation as dr;
 use file_manager_client as fmc;
-use flo_stream::MessagePublisher;
-use flo_stream::Subscriber;
 use parser::Parser;
 
-use serde::Serialize;
-use serde::Deserialize;
 
-
-
-/// ============
-/// == Errors ==
-/// ============
-
-/// Failure for missing node metadata.
-#[derive(Debug,Clone,Copy,Fail)]
-#[fail(display="Node with ID {} was not found in metadata.", _0)]
-pub struct NodeMetadataNotFound(pub ast::ID);
-
-
-
-// ==============
-// == Metadata ==
-// ==============
-
-/// Mapping between ID and metadata.
-#[derive(Debug,Clone,Default,Deserialize,Serialize)]
-pub struct Metadata {
-    /// Metadata used within ide.
-    #[serde(default="default")]
-    pub ide : IdeMetadata,
-    #[serde(flatten)]
-    /// Metadata of other users of SourceFile<Metadata> API.
-    /// Ide should not modify this part of metadata.
-    rest : serde_json::Value,
-}
-
-impl parser::api::Metadata for Metadata {}
-
-/// Metadata that belongs to ide.
-#[derive(Debug,Clone,Default,Deserialize,Serialize)]
-pub struct IdeMetadata {
-    /// Metadata that belongs to nodes.
-    node : HashMap<ast::ID,NodeMetadata>
-}
-
-/// Metadata of specific node.
-#[derive(Debug,Clone,Copy,Default,Serialize,Deserialize,Shrinkwrap)]
-pub struct NodeMetadata {
-    /// Position in x,y coordinates.
-    pub position: Option<Position>
-}
-
-/// Used for storing node position.
-#[derive(Clone,Copy,Debug,PartialEq,Serialize,Deserialize)]
-pub struct Position {
-    /// Vector storing coordinates of the visual position.
-    pub vector:Vector2<f32>
-}
-
-
-
-// ====================
-// === Module State ===
-// ====================
-
-#[derive(Debug)]
-pub struct State {
-    /// Reference to current state of module
-    source : RefCell<SourceFile<Metadata>>,
-    /// Publisher of "text changed" notifications
-    text_notifications: RefCell<notification::Publisher<notification::Text>>,
-    /// Publisher of "graph changed" notifications
-    graph_notifications: RefCell<notification::Publisher<notification::Graphs>>,
-}
-
-impl Default for State {
-    fn default() -> Self {
-        let ast = Ast::new(ast::Module{lines:default()},None);
-        Self::new(ast,default())
-    }
-}
-
-impl State {
-
-    pub fn new(ast:Ast, metadata:Metadata) -> Self {
-        State {
-            source              : RefCell::new(SourceFile{ast,metadata}),
-            text_notifications  : default(),
-            graph_notifications : default(),
-        }
-    }
-
-    pub fn update_whole_source(&self, sources:SourceFile<Metadata>) {
-        *self.source.borrow_mut() = sources;
-        self.notify(notification::Text::Invalidate,notification::Graphs::Invalidate);
-    }
-
-    pub fn source_as_string(&self) -> FallibleResult<String> {
-        Ok(String::try_from(&*self.source.borrow())?)
-    }
-
-    pub fn ast(&self) -> Ast {
-        self.source.borrow().ast.clone_ref()
-    }
-
-    /// Update current ast in module controller and emit notification about overall invalidation.
-    pub fn update_ast(&self, ast:Ast) {
-        self.source.borrow_mut().ast  = ast;
-        self.notify(notification::Text::Invalidate,notification::Graphs::Invalidate);
-    }
-
-    /// Obtains definition information for given graph id.
-    pub fn find_definition(&self,id:&dr::graph::Id) -> FallibleResult<DefinitionInfo> {
-        let module = ast::known::Module::try_new(self.source.borrow().ast.clone())?;
-        double_representation::graph::traverse_for_definition(module,id)
-    }
-
-    /// Returns metadata for given node, if present.
-    pub fn node_metadata(&self, id:ast::ID) -> FallibleResult<NodeMetadata> {
-        let data = self.source.borrow().metadata.ide.node.get(&id).cloned();
-        data.ok_or_else(|| NodeMetadataNotFound(id).into())
-    }
-
-    /// Sets metadata for given node.
-    pub fn set_node_metadata(&self, id:ast::ID, data:NodeMetadata) {
-        self.source.borrow_mut().metadata.ide.node.insert(id,data);
-        self.notify(notification::Text::Invalidate,notification::Graphs::Invalidate);
-    }
-
-    /// Removes metadata of given node and returns them.
-    pub fn take_node_metadata(&self, id:ast::ID) -> FallibleResult<NodeMetadata> {
-        let data = self.source.borrow_mut().metadata.ide.node.remove(&id);
-        data.ok_or_else(|| NodeMetadataNotFound(id).into())
-    }
-
-    pub fn subscribe_text_notifications(&self) -> Subscriber<notification::Text> {
-        self.text_notifications.borrow_mut().subscribe()
-    }
-
-    pub fn subscribe_graph_notifications(&self) -> Subscriber<notification::Graphs> {
-        self.graph_notifications.borrow_mut().subscribe()
-    }
-
-    fn notify(&self, text_change:notification::Text, graphs_change:notification::Graphs) {
-        let code_notify  = self.text_notifications.borrow_mut().publish(text_change);
-        let graph_notify = self.graph_notifications.borrow_mut().publish(graphs_change);
-        spawn(async move { futures::join!(code_notify,graph_notify); });
-    }
-
-    #[cfg(test)]
-    pub fn from_code_or_panic<S:ToString>(code:S,id_map:ast::IdMap,metadata:Metadata) -> Rc<Self> {
-        let parser = Parser::new_or_panic();
-        let ast    = parser.parse(code.to_string(),id_map).unwrap();
-        Rc::new(Self::new(ast,metadata))
-    }
-}
 
 // =======================
 // === Module Location ===
@@ -187,6 +31,7 @@ impl State {
 pub struct Location(pub Rc<String>);
 
 impl Location {
+    /// Create new location from string.
     pub fn new<S:ToString>(string:S) -> Self {
         Location(Rc::new(string.to_string()))
     }
@@ -227,7 +72,7 @@ pub struct Handle {
     /// This module's location.
     pub location: Location,
     /// The current state of module.
-    pub module: Rc<State>,
+    pub module: state::Handle,
     /// The File Manager Client handle.
     pub file_manager: fmc::Handle,
     /// The Parser handle.
@@ -240,7 +85,7 @@ impl Handle {
     /// Create a module controller for given location.
     ///
     /// It may wait for module content, because the module must initialize its state.
-    pub fn new(location:Location, module:Rc<State>, file_manager:fmc::Handle, parser:Parser)
+    pub fn new(location:Location, module:state::Handle, file_manager:fmc::Handle, parser:Parser)
     -> Self {
         let logger   = Logger::new(format!("Module Controller {}", location));
         Handle {location,module,file_manager,parser,logger}
@@ -255,7 +100,7 @@ impl Handle {
         let source = self.parser.parse_with_metadata(content)?;
         self.logger.info(|| "Code parsed");
         self.logger.trace(|| format!("The parsed ast is {:?}", source.ast));
-        self.module.update_whole_source(source);
+        self.module.update_whole(source);
         Ok(())
     }
 
@@ -319,7 +164,7 @@ impl Handle {
     ) -> FallibleResult<Self> {
         let logger   = Logger::new("Mocked Module Controller");
         let ast      = parser.parse(code.to_string(),id_map.clone())?;
-        let module   = Rc::new(State::new(ast,default()));
+        let module   = state::Handle::new(state::State::new(ast,default()));
         Ok(Handle {location,module,file_manager,parser,logger})
     }
 }
